+++ src/App.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import './styles/app.css'
-import Navbar from './components/Navbar.jsx';
-import Home from './components/Home.jsx';
-import HorseItem from './components/HorseItem.jsx';
-import Footer from './components/Footer.jsx';
-import HorsesLibrary from "./components/HorsesLibrary";
-
+import Navbar from './components/header/Navbar.jsx';
+import Home from './components/main/Home';
+import HorseItem from './components/main/HorseItem.jsx';
+import Footer from './components/footer/Footer';
+import HorsesLibrary from "./components/main/HorsesLibrary";
 
 export default function App() {
   return (
@@ -17,7 +16,7 @@ export default function App() {
           <Switch>
             <Route exact path="/" component={Home} />
             <Route exact path="/horse-library" component={HorsesLibrary} />
-            <Route path="/horse-profile/:name" component={HorseItem} />
+            <Route path="/horse-profile/:index" component={HorseItem} />
           </Switch>
           <Footer />
         </div>
