--- src/components/Footer.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export default function Footer() {
-    return (
-      <div>
-        <ul>
-          <li>Terms & Conditions</li>
-          <li>Privacy Policy</li>
-        </ul>
-      </div>
-    );
-  }
