--- src/components/Home.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function Home() {
-    return (
-      <div>
-        <h1>Need a horse? Rent a horse!</h1>
-        <p>Horses to rent!</p>
-      </div>
-    );
-  }
