--- src/components/Profile.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function Profile() {
-    return (
-      <div>
-        <h1>Profile of a horse</h1>
-      </div>
-    );
-  }
