+++ src/components/footer/Footer.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import '../../styles/footer.css';
+
+export default function Footer() {
+  return (
+    <div className='footer-container'>
+      <ul>
+        <li>Terms & Conditions</li>
+        <li>Privacy Policy</li>
+      </ul>
+    </div>
+  );
+}
