+++ src/components/main/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../../styles/home.css';
+
+export default function Home() {
+  return (
+    <div className='home-container'>
+      <h1>Need a horse? Rent a horse!</h1>
+      <p>Horses to rent!</p>
+    </div>
+  );
+}
