+++ src/styles/footer.css
@@ -1,0 +1,3 @@
+.footer-container {
+  background: rgb(212, 212, 212)
+}
