+++ src/styles/home.css
@@ -1,0 +1,3 @@
+.home-container {
+  background: rgb(212, 212, 212)
+}
