+++ src/styles/horseItem.css
@@ -1,0 +1,3 @@
+.card-container {
+  background: rgb(212, 212, 212)
+}
