+++ src/styles/horsesLibrary.css
@@ -1,0 +1,3 @@
+.library-container {
+  background: rgb(212, 212, 212);
+}
