+++ src/styles/navbar.css
@@ -1,0 +1,3 @@
+.header-container {
+  background: rgb(212, 212, 212);
+}
